Return property value as an Array of values
This was done to reduce errors in programs calling the library as they
may forget to check whether the value is an Array, assuming that it
won't be.  If this check for an Array is to be done, then this is no
different to having to extract the first or other item from the Array
by default, so it is not saving anything by not returning an Array for a
single value.